Rebase facter to 1.6.14.
This is a patch to rebase facter to 1.6.14 for rawhide (and should be a
clean change in F17).  Note I had to add a requires on rpsec (not just
rspec-core) to make this build cleanly in the %%check phase.

This patch also removes the asc file, as it is never (and has never
been) used in the process of making the facter package.

rpmlint looks reasonable to me upon build.

Bugzilla: 871211

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>