Rebase to 1.6.15 and fix issue found in bz #871211
This is a rebase to the upstream of 1.6.15. This should be a backward
compatible release with what is currently in Fedora/EPEL. The known
issue of openstack/ec2 fact exception handling has been resolved in
1.6.15.

This commit also adds the .asc file back as it used by facter
maintainers during package creation.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>