Send dmidecode errors to /dev/null in the virtual fact
When run as a non-root user, facter would always spew "/dev/mem:
Permission denied" because it called dmidecode without redirecting
stderr.

This was filed upstream as FACT-86¹.  It's unclear what is insufficient
about this method, but failing tests that are not visible aren't as much
of a concern to me as clearly useless output on every run of the
command.

¹ https://tickets.puppetlabs.com/browse/FACT-86